refactor(ui): extract AppInner state/hooks into modules
- appState.ts: AppView, YamlEditorState, UpdateFlowState types + initialYamlEditorState
- useCollectionUiPersistence.ts: last-request + expanded-folders persist hooks
- useTimelineActions.ts: timeline body load/copy/export callbacks
- useUpdateFlow.ts: update check/install state machine
- Update AppInner, AppOverlays, commands.ts, useAppKeymap to use extracted modules
