+++ src/ui/useUpdateFlow.ts
@@ -1,0 +1,138 @@
+import { useCallback, useEffect, useRef, useState } from "react"
+import type { RefObject } from "react"
+import {
+  checkForUpdates,
+  installBinaryUpdate,
+  installBrewUpdate,
+} from "../app/commands/update"
+import { showToast } from "./Toast"
+import type { UpdateFlowState } from "./appState"
+
+export function useUpdateFlow(overlayActiveRef: RefObject<boolean>) {
+  const [checkToken, setCheckToken] = useState(0)
+  const [updateFlow, setUpdateFlow] = useState<UpdateFlowState>({
+    phase: "idle",
+  })
+  const [restartVersion, setRestartVersion] = useState<string | null>(null)
+  const [updateAvailable, setUpdateAvailable] = useState<string | null>(null)
+  const updateFlowRef = useRef(updateFlow)
+  updateFlowRef.current = updateFlow
+  const installTokenRef = useRef(0)
+
+  const triggerUpdateCheck = useCallback(
+    () => setCheckToken((token) => token + 1),
+    [],
+  )
+
+  useEffect(() => {
+    if (checkToken === 0 || updateFlowRef.current.phase === "installing") return
+    let cancelled = false
+    checkForUpdates(true).then((status) => {
+      if (cancelled || overlayActiveRef.current) return
+      if (status.kind === "up_to_date") {
+        setUpdateAvailable(null)
+        showToast("Noodle is up to date", "success")
+      } else if (status.kind === "error") {
+        showToast(status.message, "error")
+      } else if (status.kind === "update_available") {
+        setUpdateFlow({
+          phase: "confirm",
+          version: status.latestVersion || "latest",
+          installType: status.installType,
+          assetUrl:
+            status.installType === "binary" ? status.assetUrl : undefined,
+          expectedSha256:
+            status.installType === "binary" ? status.expectedSha256 : undefined,
+        })
+      }
+    })
+    return () => {
+      cancelled = true
+    }
+  }, [checkToken])
+
+  useEffect(() => {
+    if (updateFlow.phase !== "installing") return
+    const update = updateFlow
+    const token = ++installTokenRef.current
+    if (update.installType === "brew") {
+      installBrewUpdate().then((result) => {
+        if (token !== installTokenRef.current) return
+        if (result.data.status === "homebrew_updated") {
+          showToast("Updated via Homebrew", "success")
+          setUpdateFlow({ phase: "done", version: update.version || "latest" })
+          setRestartVersion(update.version || "latest")
+          setUpdateAvailable(null)
+        } else {
+          const message = result.data.exit_code
+            ? `Homebrew upgrade failed (exit ${result.data.exit_code})`
+            : "Homebrew upgrade failed"
+          showToast(message, "error")
+          setUpdateFlow({ phase: "failed", message })
+        }
+      })
+      return
+    }
+    if (
+      update.installType === "binary" &&
+      update.assetUrl &&
+      update.expectedSha256
+    ) {
+      installBinaryUpdate(
+        update.version,
+        update.assetUrl,
+        update.expectedSha256,
+      ).then((result) => {
+        if (token !== installTokenRef.current) return
+        if (result.data.status === "updated") {
+          const version = result.data.version ?? update.version
+          showToast(`Updated to ${version}`, "success")
+          setUpdateFlow({ phase: "done", version })
+          setRestartVersion(version)
+          setUpdateAvailable(null)
+        } else {
+          const message =
+            (result.data as Record<string, string>).reason ?? "Update failed"
+          showToast(message, "error")
+          setUpdateFlow({ phase: "failed", message })
+        }
+      })
+      return
+    }
+    setUpdateFlow({ phase: "idle" })
+  }, [updateFlow])
+
+  useEffect(() => {
+    let cancelled = false
+    checkForUpdates().then((status) => {
+      if (!cancelled && status.kind === "update_available") {
+        setUpdateAvailable(status.latestVersion)
+      }
+    })
+    return () => {
+      cancelled = true
+    }
+  }, [])
+
+  const confirmInstall = useCallback(() => {
+    const update = updateFlowRef.current
+    if (update.phase !== "confirm") return
+    setUpdateFlow({
+      phase: "installing",
+      version: update.version,
+      installType: update.installType,
+      assetUrl: update.assetUrl,
+      expectedSha256: update.expectedSha256,
+    })
+  }, [])
+  const cancelUpdate = useCallback(() => setUpdateFlow({ phase: "idle" }), [])
+
+  return {
+    updateFlow,
+    restartVersion,
+    updateAvailable,
+    triggerUpdateCheck,
+    confirmInstall,
+    cancelUpdate,
+  }
+}
